refactored llmparse, added task executor module,  some bug fixes
